tests/factories.py
import factory
from kalibro_client.processor import Project

class ProjectFactory(factory.Factory):
    class Meta:
        model = Project

    name = 'A random Project'
    description = 'A real example Project'